Restructure the repository according to OpenStack
This patch is restructuring the VirtualBMC repository according to the
OpenStack template for projects (Cookiecutter).

This patch also removes the dependency on the python-daemon library
since it is not present in the OpenStack global-requirements, now a new
class called "detach_process" has been added to the virtualbmc/utils.py
module that will take care of detaching the process context from its
parent and session.

Change-Id: Id4ae1db5b73a18abc54276fe1dfbf3ceff7adc06