Add support for a "global" configuration file
Prior to this patch, virutalbmc could only be configured in a per-user
basis. The configuration file as well the path where the VBMCs are going
to be created always lived in the users home directory.

This patch is adding support for a "global" configuration file which
will live in /etc/virtualbmc/virtualbmc.conf. When vbmc does not find a
configuration file in the user directory it will then fallback to the
global (if it exist, otherwise it will just uses the defaults values).

Also, a new configuration option called "config_dir" has been added.
This allows users to configure whatever directory they want to save the
created VirtualBMCs where before it always pointed to ~/.vbmc/config.

Closes-Bug: #1647341
Change-Id: Id8bca620d1543436f32ecc650fc28cbc3f13f373