Return proper errors on BMC action failures
Prior to this patch VirtualBMC methods would return None on most on most
failures and that would get translated into return code 0 by pyghmi. So,
even when a node failed to stop, start, get or set the boot device the
error was never propagated to the IPMI client.

This patch changes such behavior to include proper errors upon failures.

Change-Id: If761f8b56c16d265c5e6917e4ee7352126b0c90e
Closes-Bug: #1647229