Clean up zuul files
This moves the irrelevant-files section from project.yaml to
legacy-virtualbmc-jobs.yaml. It adds more irrelevant files to
this list, and changes it so that requirements.yaml is now
relevant.

It removes two projects that are required projects in the
parent 'legacy-dsvm-base'.

The run, post-run, and timeout attributes were moved to the
bottom to be more consistent with how it is done in other
ironic-related projects.

Change-Id: Ib4a88d0589a0bfd85713918c7cdfe212c0e54645