Fix stop command by using default SIGTERM handler in children
The daemon child processes do not respond to a SIGTERM. This means that
they stay running indefinitely. This is because the manager process
installs a SIGTERM handler to propagate the signal to children, and this
handler is inherited by the children.

Return children to the default handler for SIGTERM.

Change-Id: I784e7cd9e0545a591150dc21a8232bd85672c21c
Story: 2003534
Task: 24819